consolidated all the graphics test into a single source file, updated makefile
