--- headers/test_suit.h
@@ -1,22 +1,0 @@
-#ifndef _TEST_SUITE_H_
-#define _TEST_SUITE_H_
-
-#include <stdio.h>
-#include <assert.h>
-#include <stdlib.h>
-#include "SDL2/SDL.h"
-#include "init.h"
-#include "input.h"
-#include "display.h"
-#include "gameobjects.h"
-
-int cleantest();
-int test1();
-int test2();
-int test3();
-int test4();
-int test5();
-int test6();
-int test7();
-
-#endif
